+++ src/pages/index/threads.vue
@@ -59,7 +59,7 @@
             </ulink>
             <div class="count">
               <p class="title">
-                <i18n en="Latest" zh="时新" />
+                <i18n en="Latest" zh="时时" />
               </p>
               <p class="secondary">
                 <i18n en="threads" zh="清静念" />