+++ src/pages/snippets/shared.ts
@@ -1,13 +1,6 @@
-import { ref, shallowReactive, shallowRef, computed } from 'vue'
 import { useEnhancer } from '/@/app/enhancer'
-import { useZhihuLatestAnswersStore } from '/@/stores/media'
-import type { ZhihuAnswerItem, ZhihuAnswersResponse } from '/@/server/getters/zhihu'
 import { firstUpperCase } from '/@/transforms/text'
 import { Language, LanguageKey } from '/@/language'
-import { TunnelModule } from '/@/constants/tunnel'
-import { delayPromise } from '/@/utils/delayer'
-import { isClient } from '/@/app/environment'
-import tunnel from '/@/services/tunnel'
 import { META } from '/@/config/app.config'
 
 export const i18nTitle = {
@@ -26,42 +19,4 @@ export const useSnippetsPageMeta = () => {
       description: description
     }
   })
-}
-
-export const useSnippetsPageData = () => {
-  const zhihuLatestAnswers = useZhihuLatestAnswersStore()
-  const loading = ref(false)
-  const localAnswers = shallowReactive<Array<ZhihuAnswerItem>>([])
-  const lastPaging = shallowRef<ZhihuAnswersResponse['paging'] | null>(null)
-  const nextPageOffset = computed(() => {
-    const paging = lastPaging.value ?? zhihuLatestAnswers.data?.paging
-    return new URL(paging!.next)?.searchParams.get('offset')
-  })
-  const finished = computed(() => lastPaging.value?.is_end ?? zhihuLatestAnswers.data?.paging.is_end ?? false)
-  const allAnswers = computed(() => {
-    const latestMedias = zhihuLatestAnswers.data?.data ?? []
-    return [...latestMedias, ...localAnswers]
-  })
-
-  const fetchMoreAnswers = async () => {
-    try {
-      loading.value = true
-      const request = tunnel.dispatch<ZhihuAnswersResponse>(TunnelModule.ZhihuAnswers, {
-        offset: nextPageOffset.value
-      })
-      const response = await (isClient ? delayPromise(360, request) : request)
-      localAnswers.push(...response.data)
-      lastPaging.value = response.paging
-    } finally {
-      loading.value = false
-    }
-  }
-
-  return {
-    loading,
-    finished,
-    allAnswers,
-    fetchMoreAnswers,
-    zhihuLatestAnswers
-  }
 }